requirements.txt
Pillow==2.4.0
enum==0.4.4
wsgiref==0.1.2
